+++ game_server/Application.cpp
@@ -11,6 +11,7 @@
 #include "handlers/PacketHandlers.hpp"
 #include "network/Connection.hpp"
 #include "network/packets/server/ClientDisconnect.hpp"
+#include "utils/Chrono.hpp"
 
 #include <l2cpp/Exception.hpp>
 #include <l2cpp/Misc.hpp>
@@ -69,18 +70,6 @@ catch (l2cpp::Exception const & e)
     return false;
 }
 
-struct Clock
-{
-    std::chrono::steady_clock::time_point start;
-    Clock(): start(std::chrono::steady_clock::now()) {}
-    auto restart() -> std::chrono::steady_clock::duration
-    {
-        auto const d = std::chrono::steady_clock::now() - start;
-        start = std::chrono::steady_clock::now();
-        return d;
-    }
-};
-
 bool Application::ApplicationImpl::run()
 {
     signalSet.async_wait([this] (auto const & ec, int s) { onSignal(ec, s); });
@@ -93,7 +82,7 @@ bool Application::ApplicationImpl::run()
     SPDLOG_INFO("Listening for clients on {}:{}", ip, port);
 
     SPDLOG_INFO("Server running. Input CTRL+C to initiate shutdown…");
-    Clock loopClock, worldClock, ioClock;
+    Utils::Chrono::Clock loopClock, worldClock, ioClock;
     while (!ioContext.stopped())
     {
         ioClock.restart();