+++ game_server/game/constants/ActionType.hpp
@@ -1,0 +1,13 @@
+/// @author    Chnossos
+/// @date      Created on 2026-03-28
+
+#pragma once
+
+// Project includes
+#include <l2cpp/Typedefs.hpp>
+
+enum class ActionType
+{
+    Move,
+    Attack,
+};